chore: refactor code [Finishes #165279323]
