+++ server/middleware/validators/signUpValidator.js
@@ -1,4 +1,5 @@
-import { check, validationResult } from 'express-validator/check';
+import { check } from 'express-validator/check';
+import errs from './errorMessages';
 
 const userInput = [
   check('email').isEmail().trim().withMessage('input a valid email address'),
@@ -9,21 +10,7 @@ const userInput = [
   check('lastName').not().isEmpty().withMessage('lastName cannot be empty'),
   check('firstName').isAlpha().trim().withMessage('firstName can ony contain letters'),
   check('lastName').isAlpha().trim().withMessage('lastName can ony contain letters'),
-  (req, res, next) => {
-    const messages = [];
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      errors.array().forEach((err) => {
-        messages.push(err.msg);
-      });
-
-      return res.status(422).json({
-        status: 422,
-        errors: messages,
-      });
-    }
-    return next();
-  },
+  errs.displayErrs,
 ];
 
 export default userInput;
